+++ internal/provider/model_meraki_network_floor_plan.go
@@ -35,7 +35,7 @@ import (
 type NetworkFloorPlan struct {
 	Id                   types.String  `tfsdk:"id"`
 	NetworkId            types.String  `tfsdk:"network_id"`
-	FloorNumber          types.Int64   `tfsdk:"floor_number"`
+	FloorNumber          types.Float64 `tfsdk:"floor_number"`
 	ImageContents        types.String  `tfsdk:"image_contents"`
 	Name                 types.String  `tfsdk:"name"`
 	BottomLeftCornerLat  types.Float64 `tfsdk:"bottom_left_corner_lat"`
@@ -65,7 +65,7 @@ func (data NetworkFloorPlan) getPath() string {
 func (data NetworkFloorPlan) toBody(ctx context.Context, state NetworkFloorPlan) string {
 	body := ""
 	if !data.FloorNumber.IsNull() {
-		body, _ = sjson.Set(body, "floorNumber", data.FloorNumber.ValueInt64())
+		body, _ = sjson.Set(body, "floorNumber", data.FloorNumber.ValueFloat64())
 	}
 	if !data.ImageContents.IsNull() {
 		body, _ = sjson.Set(body, "imageContents", data.ImageContents.ValueString())
@@ -112,9 +112,9 @@ func (data NetworkFloorPlan) toBody(ctx context.Context, state NetworkFloorPlan)
 
 func (data *NetworkFloorPlan) fromBody(ctx context.Context, res meraki.Res) {
 	if value := res.Get("floorNumber"); value.Exists() && value.Value() != nil {
-		data.FloorNumber = types.Int64Value(value.Int())
+		data.FloorNumber = types.Float64Value(value.Float())
 	} else {
-		data.FloorNumber = types.Int64Null()
+		data.FloorNumber = types.Float64Null()
 	}
 	if value := res.Get("name"); value.Exists() && value.Value() != nil {
 		data.Name = types.StringValue(value.String())
@@ -133,9 +133,9 @@ func (data *NetworkFloorPlan) fromBody(ctx context.Context, res meraki.Res) {
 // "managed" elements, instead of all elements.
 func (data *NetworkFloorPlan) fromBodyPartial(ctx context.Context, res meraki.Res) {
 	if value := res.Get("floorNumber"); value.Exists() && !data.FloorNumber.IsNull() {
-		data.FloorNumber = types.Int64Value(value.Int())
+		data.FloorNumber = types.Float64Value(value.Float())
 	} else {
-		data.FloorNumber = types.Int64Null()
+		data.FloorNumber = types.Float64Null()
 	}
 	if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
 		data.Name = types.StringValue(value.String())